requirements.txt
flask
gunicorn
face_recognition
opencv-python
pillow
numpy
matplotlib